+++ src/player/view.c
@@ -25,7 +25,7 @@ float SV_CalcRoll (vec3_t angles, vec3_t velocity)
 	
 	side = DotProduct (velocity, right);
 	sign = side < 0 ? -1 : 1;
-	side = fabs(side);
+	side = fabsf(side);
 	
 	value = sv_rollangle->value;
 
@@ -1164,7 +1164,7 @@ void ClientEndServerFrame (edict_t *ent)
 		bobtime *= 4;
 
 	bobcycle = (int)bobtime;
-	bobfracsin = fabs(sin(bobtime*M_PI));
+	bobfracsin = fabsf(sin(bobtime*M_PI));
 
 	// detect hitting the floor
 	P_FallingDamage (ent);
